Adds WordPress.org assets
These existed in the previous SVN repository, but are not a part
of the Secretary plugin itself. Luckily, the `WordPress.org Plugin
Deploy` GitHub Action has its own way of handling this, by putting
such assets in the '.wordpress-org' folder, which it explicitly
avoids publishing as part of the plugin.