Add .distignore file
We need to explicitly declare README.md as a file that should not
be committed to the SVN repository (it is intended only for
dev/contributing purposes, to be displayed on GitHub.com, not on
the WordPress site).

If we don't exclude such files, then the GitHub action for updating
'tested up to' will fail, as it will detect that more than just
the readme.txt/assets have changed.