Delete 'assets/' folder
This folder was (accidentally?) included in previous versions of
Secretary, when the assets in question are only used on WordPress.org
for marketing reasons.
I added them to the GitHub repo to have the 'action-wordpress-plugin-asset-update'
GitHub Action stop complaining about their omission (they were in
the SVN repo but not the Git repo originally).